Split csvs.
