+++ util/test/test_allocations.py
@@ -1,6 +1,6 @@
 from enforce_typing import enforce_types
 
-from util import csvs
+from util.volume import csvs
 from util.volume.allocations import allocsToStakes, loadStakes
 
 # for shorter lines
