+++ util/test/test_dftool_noganache.py
@@ -2,10 +2,11 @@ import os
 import subprocess
 from typing import List
 
-from enforce_typing import enforce_types
 import pytest
+from enforce_typing import enforce_types
 
-from util import csvs, dftool_module
+from util import dftool_module
+from util.volume import csvs
 
 
 @enforce_types
